move cat button
