+++ scripts/screens/EventsScreen.py
@@ -21,8 +21,8 @@ from scripts.game_structure.screen_settings import MANAGER
 from scripts.game_structure.ui_elements import (
     UIModifiedScrollingContainer,
     IDImageButton,
-    CatButton,
 )
+from scripts.ui.elements.cat_button import CatButton
 from scripts.ui.elements.surface_image_button import UISurfaceImageButton
 from scripts.screens.screens_core.screens_core import rebuild_moon_n_season_indicator
 from scripts.ui.elements.save_button import UISaveButton
